pull from Microsoft:master (#1)
* Reducing number of submodules pulled by boost and depth of git clone

* Correcting build times in build requirements section of README (#27)

* Correcting build times in build requirements section of README

Adding correct times, previously they all said 2 hours.

* Corrected 1 hours to 1 hour

* Changes to clarify and highlight the additional disk space needed on azure

Previously the note was easy to miss.

* Revert "Reducing number of submodules pulled by boost and depth of git clone"

* Updating link in README.md to point to LICENSE (#34)

Fixed a link in README.md to point to LICENSE instead of LICENSE.txt

* Adding docs for deploying local cluster from packages generated in the build

* Decrease docs folder readme font header size

* Update based on comments

* Adding a basic end-to-end test using a onebox cluster (#15)

* Adding a basic end-to-end test using a onebox cluster

Added a script to launch and interact with a onebox cluster in a Docker container.
Added a test that downloads a sample application and deploys it to the cluster
inside of a Docker container. The test passes if the application http endpoint can
be hit.